git-tfs-id: [http://devvm01tfs03.telamericamedia.com:8080/tfs/CMWPA]$/Telamerica/Broadcast/OnDeck;C20839

+++ Source/Tam/Maestro/Services/Broadcast/Services.Broadcast/Converters/RateImport/AssemblyRateImporter.cs
@@ -199,7 +199,7 @@ namespace Services.Broadcast.Converters.RateImport
                 var flightWeekAudience = new StationProgramFlightWeekAudience()
                 {
                     Audience = _audienceMapping[audienceField.Key],
-                    Impressions = float.Parse(currentData[audienceField.Value]) * 1000
+                    Impressions = double.Parse(currentData[audienceField.Value]) * 1000
                 };
                 flightWeekAudiences.Add(flightWeekAudience);
             }
